refactor(undo): extract undo/redo into a dedicated package
Move the undo/redo code out of the top-level managers directory into a
`managers/undo/` package and split the manager, which had grown to own the
stacks, replay, the recording state machine, and the request handlers all at
once:

- `undo/undo.py`: the domain-facing vocabulary (entries, recorders, batch,
  replay helpers).
- `undo/recording.py`: the recording state machine (`RecordingSession` plus its
  framing dataclasses), decoupled from the stacks via callbacks.
- `undo/manager.py`: a slim `UndoManager` owning the stacks, replay, and request
  handlers, delegating recording to a `RecordingSession`.
- `undo/node.py`: the node-domain recorders (was `node_manager_undo.py`).

No behavior change.
