+++ src/griptape_nodes/retained_mode/managers/undo/__init__.py
@@ -1,0 +1,35 @@
+"""Undo/redo subsystem.
+
+Public surface:
+    - `UndoManager`: the mechanism (undo/redo stacks, replay, request handlers, recording API).
+    - The vocabulary domains use to describe reversals: `UndoEntry`, `UndoBatch`,
+      `RequestReplayUndoEntry`, `UndoRecorder`, `RecorderCapture`, `UndoEntryReplayError`, and the
+      `dispatch_expecting` / `dispatch_expecting_success` replay helpers.
+
+Per-domain reversal knowledge lives in sibling modules (e.g. `undo.node`) and is registered with
+the manager by the owning domain manager.
+"""
+
+from griptape_nodes.retained_mode.managers.undo.manager import UndoManager
+from griptape_nodes.retained_mode.managers.undo.undo import (
+    RecorderCapture,
+    RequestReplayUndoEntry,
+    UndoBatch,
+    UndoEntry,
+    UndoEntryReplayError,
+    UndoRecorder,
+    dispatch_expecting,
+    dispatch_expecting_success,
+)
+
+__all__ = [
+    "RecorderCapture",
+    "RequestReplayUndoEntry",
+    "UndoBatch",
+    "UndoEntry",
+    "UndoEntryReplayError",
+    "UndoManager",
+    "UndoRecorder",
+    "dispatch_expecting",
+    "dispatch_expecting_success",
+]
